requirements.txt
# Core application dependencies
apscheduler
beautifulsoup4
feedparser
google-generativeai
httpx[http2]
lxml
Pillow
python-dateutil
python-dotenv
python-slugify
requests
trafilatura

# Dashboard dependencies
Flask
psutil